+++ pytest_fixtures/core/contenthosts.py
@@ -10,7 +10,6 @@ from broker import Broker
 from robottelo import constants
 from robottelo.config import settings
 from robottelo.hosts import ContentHost
-from robottelo.hosts import get_sat_rhel_version
 from robottelo.hosts import Satellite
 
 
@@ -221,48 +220,6 @@ def oracle_host(request, version):
     }
     with Broker(**host_conf(request), host_class=ContentHost) as host:
         yield host
-
-
-def get_deploy_args(request):
-    deploy_args = {
-        'deploy_rhel_version': get_sat_rhel_version().base_version,
-        'deploy_flavor': settings.flavors.default,
-        'promtail_config_template_file': 'config_sat.j2',
-        'workflow': 'deploy-rhel',
-    }
-    if hasattr(request, 'param'):
-        if isinstance(request.param, dict):
-            deploy_args.update(request.param)
-        else:
-            deploy_args['deploy_rhel_version'] = request.param
-    return deploy_args
-
-
-@pytest.fixture
-def sat_ready_rhel(request):
-    deploy_args = get_deploy_args(request)
-    with Broker(**deploy_args, host_class=Satellite) as host:
-        yield host
-
-
-@pytest.fixture(scope='module')
-def module_sat_ready_rhels(request):
-    deploy_args = get_deploy_args(request)
-    with Broker(**deploy_args, host_class=Satellite, _count=2) as hosts:
-        yield hosts
-
-
-@pytest.fixture
-def cap_ready_rhel():
-    rhel8 = settings.content_host.rhel8.vm
-    deploy_args = {
-        'deploy_rhel_version': rhel8.deploy_rhel_version,
-        'deploy_flavor': 'satqe-ssd.standard.std',
-        'workflow': rhel8.workflow,
-    }
-    with Broker(**deploy_args, host_class=Capsule) as host:
-        yield host
-
 
 
 @pytest.fixture(scope='module', params=[{'rhel_version': 8, 'no_containers': True}])
